+++ CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 ## Unreleased
 
 - Improved support for using custom program sources. [#741](https://github.com/pulumi/pulumi-kubernetes-operator/pull/741) 
+- Improved Status logging. [#742](https://github.com/pulumi/pulumi-kubernetes-operator/pull/742)
 
 ## 2.0.0-beta.1 (2024-10-18)